--- crates/rapier-urdf/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "rapier-urdf"
-version = "0.1.0"
-edition = "2021"
-
-[features]
-stl = ["rapier-stl"]
-
-[dependencies]
-log = "0.4"
-anyhow = "1"
-bitflags = "2"
-# NOTE: we are not using the (more recent) urdf-rs crate because of https://github.com/openrr/urdf-rs/issues/94
-xurdf = "0.2"
-
-rapier3d = { versions = "0.19", path = "../rapier3d" }
-rapier-stl = { version = "0.1.0", path = "../rapier-stl", optional = true }
