+++ crates/rapier2d/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "rapier2d"
 version = "0.19.0"
-authors = ["Sébastien Crozet <developer@crozet.re>"]
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
 description = "2-dimensional physics engine in Rust."
 documentation = "https://docs.rs/rapier2d"
 homepage = "https://rapier.rs"
