+++ crates/rapier3d-f64/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "rapier3d-f64"
 version = "0.19.0"
-authors = ["Sébastien Crozet <developer@crozet.re>"]
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
 description = "3-dimensional physics engine in Rust."
 documentation = "https://docs.rs/rapier3d"
 homepage = "https://rapier.rs"
