+++ crates/rapier3d-stl/CHANGELOG.md
@@ -1,0 +1,9 @@
+## Unreleased
+
+This is the initial release of the `rapier3d-stl` crate.
+
+### Added
+
+- Add `load_from_path` for creating a shape from a stl file.
+- Add `load_from_reader` for creating a shape from an object implementing `Read`.
+- Add `load_from_raw_mesh` for creating a shape from an already loaded `IndexedMesh`.
