+++ crates/rapier3d-stl/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "rapier3d-stl"
+version = "0.1.0"
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
+description = "STL file loader for the 3D rapier physics engine."
+documentation = "https://docs.rs/rapier3d-stl"
+homepage = "https://rapier.rs"
+repository = "https://github.com/dimforge/rapier"
+readme = "README.md"
+categories = ["science", "game-development", "mathematics", "simulation", "wasm"]
+keywords = ["physics", "joints", "multibody", "robotics", "urdf"]
+license = "Apache-2.0"
+edition = "2021"
+
+[dependencies]
+thiserror = "1.0.61"
+stl_io = "0.7"
+
+rapier3d = { version = "0.19", path = "../rapier3d" }
