+++ crates/rapier3d-stl/README.md
@@ -1,0 +1,18 @@
+## STL loader for the Rapier physics engine
+
+Rapier is a set of 2D and 3D physics engines for games, animation, and robotics. The `rapier3d-stl`
+crate lets you create a shape compatible with `rapier3d` and `parry3d` (the underlying collision-detection
+library) from an STL file.
+
+## Resources and discussions
+
+- [Dimforge](https://dimforge.com): See all the open-source projects we are working on! Follow our announcements
+  on our [blog](https://www.dimforge.com/blog).
+- [User guide](https://www.rapier.rs/docs/): Learn to use Rapier in your project by reading the official User Guides.
+- [Discord](https://discord.gg/vt9DJSW): Come chat with us, get help, suggest features, on Discord!
+- [NPM packages](https://www.npmjs.com/search?q=%40dimforge): Check out our NPM packages for Rapier, if you need to
+  use it with JavaScript/Typescript.
+
+Please make sure to familiarize yourself with our [Code of Conduct](CODE_OF_CONDUCT.md)
+and our [Contribution Guidelines](CONTRIBUTING.md) before contributing or participating in
+discussions with the community.
