+++ crates/rapier3d-urdf/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "rapier3d-urdf"
+version = "0.1.0"
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
+description = "URDF file loader for the 3D rapier physics engine."
+documentation = "https://docs.rs/rapier3d-urdf"
+homepage = "https://rapier.rs"
+repository = "https://github.com/dimforge/rapier"
+readme = "README.md"
+categories = ["science", "game-development", "mathematics", "simulation", "wasm"]
+keywords = ["physics", "joints", "multibody", "robotics", "urdf"]
+license = "Apache-2.0"
+edition = "2021"
+
+[features]
+stl = ["rapier3d-stl"]
+
+[dependencies]
+log = "0.4"
+anyhow = "1"
+bitflags = "2"
+# NOTE: we are not using the (more recent) urdf-rs crate because of https://github.com/openrr/urdf-rs/issues/94
+xurdf = "0.2"
+
+rapier3d = { version = "0.19", path = "../rapier3d" }
+rapier3d-stl = { version = "0.1.0", path = "../rapier3d-stl", optional = true }
