+++ crates/rapier_testbed3d-f64/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "rapier_testbed3d-f64"
 version = "0.19.0"
-authors = ["Sébastien Crozet <developer@crozet.re>"]
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
 description = "Testbed for the Rapier 3-dimensional physics engine in Rust."
 homepage = "http://rapier.org"
 repository = "https://github.com/dimforge/rapier"
