+++ examples2d/all_examples2.rs
@@ -56,7 +56,7 @@ fn demo_name_from_command_line() -> Option<String> {
     None
 }
 
-#[cfg(any(target_arch = "wasm32", target_arch = "asmjs"))]
+#[cfg(any(target_arch = "wasm32"))]
 fn demo_name_from_url() -> Option<String> {
     None
     //    let window = stdweb::web::window();
@@ -64,7 +64,7 @@ fn demo_name_from_url() -> Option<String> {
     //    Some(hash[1..].to_string())
 }
 
-#[cfg(not(any(target_arch = "wasm32", target_arch = "asmjs")))]
+#[cfg(not(any(target_arch = "wasm32")))]
 fn demo_name_from_url() -> Option<String> {
     None
 }
