+++ examples3d-f64/Cargo.toml
@@ -1,22 +1,22 @@
 [package]
-name    = "rapier-examples-3d-f64"
+name = "rapier-examples-3d-f64"
 version = "0.1.0"
-authors = [ "Sébastien Crozet <developer@crozet.re>" ]
+authors = ["Sébastien Crozet <sebcrozet@dimforge.com>"]
 edition = "2021"
 default-run = "all_examples3-f64"
 
 [features]
-parallel = [ "rapier3d-f64/parallel", "rapier_testbed3d-f64/parallel" ]
-simd-stable = [ "rapier3d-f64/simd-stable" ]
-simd-nightly = [ "rapier3d-f64/simd-nightly" ]
-enhanced-determinism = [ "rapier3d-f64/enhanced-determinism" ]
+parallel = ["rapier3d-f64/parallel", "rapier_testbed3d-f64/parallel"]
+simd-stable = ["rapier3d-f64/simd-stable"]
+simd-nightly = ["rapier3d-f64/simd-nightly"]
+enhanced-determinism = ["rapier3d-f64/enhanced-determinism"]
 
 [dependencies]
 rand = "0.8"
-getrandom = { version = "0.2", features = [ "js" ] } 
-Inflector  = "0.11"
+getrandom = { version = "0.2", features = ["js"] }
+Inflector = "0.11"
 wasm-bindgen = "0.2"
-obj-rs = { version =  "0.7", default-features = false }
+obj-rs = { version = "0.7", default-features = false }
 bincode = "1"
 serde = "1"
